@@ -73,7 +73,7 @@
     "ts-loader": "7.0.5",
     "ts-node": "8.10.2",
     "tslint": "5.18.0",
-    "typescript": "3.8.3",
+    "typescript": "3.9.6",
     "webpack": "4.43.0",
     "webpack-cli": "3.3.12"
   },
