@@ -20,58 +20,58 @@
   "private": true,
   "license": "See in LICENSE",
   "devDependencies": {
-    "@angular/animations": "9.1.11",
-    "@angular-builders/custom-webpack": "9.2.0",
-    "@angular/cli": "9.1.10",
-    "@angular/common": "9.1.11",
-    "@angular/compiler": "9.1.11",
-    "@angular/compiler-cli": "9.1.11",
-    "@angular/core": "9.1.11",
-    "@angular/localize": "9.1.11",
-    "@angular-devkit/build-angular": "0.901.10",
-    "@angular-devkit/build-optimizer": "0.901.10",
-    "@angular-devkit/core": "9.1.10",
-    "@angular/forms": "9.1.11",
-    "@angular/language-service": "9.1.11",
-    "@angular/platform-browser": "9.1.11",
-    "@angular/platform-browser-dynamic": "9.1.11",
-    "@angular/router": "9.1.11",
-    "@ng-bootstrap/ng-bootstrap": "6.1.0",
-    "@ngtools/webpack": "9.1.10",
+    "@angular/animations": "10.0.3",
+    "@angular-builders/custom-webpack": "10.0.0-beta.0",
+    "@angular/cli": "10.0.2",
+    "@angular/common": "10.0.3",
+    "@angular/compiler": "10.0.3",
+    "@angular/compiler-cli": "10.0.3",
+    "@angular/core": "10.0.3",
+    "@angular/localize": "10.0.3",
+    "@angular-devkit/build-angular": "0.1000.2",
+    "@angular-devkit/build-optimizer": "0.1000.2",
+    "@angular-devkit/core": "10.0.2",
+    "@angular/forms": "10.0.3",
+    "@angular/language-service": "10.0.3",
+    "@angular/platform-browser": "10.0.3",
+    "@angular/platform-browser-dynamic": "10.0.3",
+    "@angular/router": "10.0.3",
+    "@ng-bootstrap/ng-bootstrap": "7.0.0",
+    "@ngtools/webpack": "10.0.2",
     "@ngrx/effects": "9.2.0",
     "@ngrx/router-store": "9.2.0",
     "@ngrx/store": "9.2.0",
     "@ngrx/store-devtools": "9.2.0",
     "@types/node": "12.12.29",
-    "angular-confirmation-popover": "4.2.2",
+    "angular-confirmation-popover": "6.0.0",
     "angular-notifier": "6.0.1",
     "bootstrap": "3.4.1",
-    "codelyzer": "5.2.1",
+    "codelyzer": "6.0.0",
     "dragula": "3.7.2",
-    "@fortawesome/angular-fontawesome": "0.6.1",
-    "@fortawesome/fontawesome-svg-core": "1.2.28",
-    "@fortawesome/free-solid-svg-icons": "5.13.0",
-    "@fortawesome/free-regular-svg-icons": "5.13.0",
-    "jasmine-core": "3.3.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "3.1.4",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "2.0.6",
-    "karma-jasmine": "1.1.2",
-    "karma-jasmine-html-reporter": "1.3.1",
+    "@fortawesome/angular-fontawesome": "0.7.0",
+    "@fortawesome/fontawesome-svg-core": "1.2.29",
+    "@fortawesome/free-solid-svg-icons": "5.13.1",
+    "@fortawesome/free-regular-svg-icons": "5.13.1",
+    "jasmine-core": "3.5.0",
+    "jasmine-spec-reporter": "5.0.2",
+    "karma": "5.1.0",
+    "karma-chrome-launcher": "3.1.0",
+    "karma-coverage-istanbul-reporter": "3.0.3",
+    "karma-jasmine": "3.3.1",
+    "karma-jasmine-html-reporter": "1.5.4",
     "ng2-dragula": "2.1.1",
     "ng2-nouislider": "1.8.2",
-    "ngx-clipboard": "13.0.0",
-    "ngx-select-ex": "3.7.6",
+    "ngx-clipboard": "13.0.1",
+    "ngx-select-ex": "4.0.0",
     "ngrx-store-freeze": "0.2.4",
-    "nouislider": "14.1.1",
-    "protractor": "5.4.0",
-    "rxjs": "6.5.4",
+    "nouislider": "14.6.0",
+    "protractor": "7.0.0",
+    "rxjs": "6.6.0",
     "semver": "7.1.1",
     "ts-keycode-enum": "^1.0.6",
     "uhk-common": "1.0.0",
     "xml-loader": "1.2.1",
-    "zone.js": "0.10.2"
+    "zone.js": "0.10.3"
   },
   "dependencies": {
     "classlist.js": "1.1.20150312",
