@@ -1,8 +1,11 @@
-import { EventEmitter, Output } from '@angular/core';
+import { Component, EventEmitter, Output } from '@angular/core';
 import { KeyAction } from 'uhk-common';
 
 import { RemapInfo } from '../../../models/remap-info';
 
+@Component({
+    template: ''
+})
 export abstract class Tab {
     @Output() validAction = new EventEmitter<boolean>();
 
